Correcting color struct

+++ GameServerCore/Content/Color.cs
@@ -5,6 +5,6 @@
         public byte R;
         public byte G;
         public byte B;
-        public byte A;
+        public float A;
     }
 }
